omnihub-function/requirements.txt
#cloud function
functions-framework==3.*
google-cloud-storage
google-cloud-aiplatform